Rename things and add docs
