+++ src/Sharpliner.Model/AzureDevOps/AzureDevOpsPipelineDefinition.cs
@@ -1,0 +1,8 @@
+﻿using Sharpliner.Model.Definition;
+
+namespace Sharpliner.Model.AzureDevOps
+{
+    public abstract class AzureDevOpsPipelineDefinition : PipelineDefinitionBase
+    {
+    }
+}
