+++ src/Sharpliner.Model/AzureDevOps/Job.cs
@@ -25,7 +25,7 @@ namespace Sharpliner.Model.AzureDevOps
 
         public TimeSpan? CancelTimeout { get; init; }
 
-        public List<Variable> Variables { get; init; } = new();
+        public List<SingleVariable> Variables { get; init; } = new();
 
         public List<Step> Steps { get; init; } = new();
