+++ src/Sharpliner.Model/AzureDevOps/Stage.cs
@@ -12,7 +12,7 @@ namespace Sharpliner.Model.AzureDevOps
 
         public string? Condition { get; init; }
 
-        public List<Variable> Variables { get; init; } = new();
+        public List<SingleVariable> Variables { get; init; } = new();
 
         public List<Job> Jobs { get; init; } = new();
