+++ src/Sharpliner.Model/GHActions/GitHubActionsPipelineDefinition.cs
@@ -1,0 +1,8 @@
+﻿using Sharpliner.Model.Definition;
+
+namespace Sharpliner.Model.GHActions
+{
+    public abstract class GitHubActionsPipelineDefinition : PipelineDefinitionBase
+    {
+    }
+}
